feat: add forge fmt option
