+++ client/src/formatter/index.ts
@@ -1,102 +1,25 @@
-import * as prettier from "prettier";
 import * as vscode from "vscode";
-import * as path from "path";
-import * as prettierPluginSolidity from "prettier-plugin-solidity";
+import * as prettier from "./prettierFormatter";
+import * as forge from "./forgeFormatter";
 
-export function formatDocument(
+export async function formatDocument(
   document: vscode.TextDocument,
   context: vscode.ExtensionContext
-): vscode.TextEdit[] {
-  const rootPath = getCurrentWorkspaceRootFsPath();
-
-  if (rootPath === undefined) {
-    return [];
-  }
-
-  const ignoreOptions = {
-    ignorePath: path.join(rootPath, ".prettierignore"),
-  };
-
-  const fileInfo = prettier.getFileInfo.sync(
-    document.uri.fsPath,
-    ignoreOptions
-  );
-
-  if (fileInfo.ignored) {
-    return [];
-  }
-
-  const source = document.getText();
-
-  const options = {
-    useCache: false,
-    parser: "solidity-parse",
-    pluginSearchDirs: [context.extensionPath],
-    plugins: [prettierPluginSolidity],
-  };
-
-  const config =
-    prettier.resolveConfig.sync(document.uri.fsPath, options) ??
-    defaultConfig();
-
-  Object.assign(options, config);
-
-  const firstLine = document.lineAt(0);
-  const lastLine = document.lineAt(document.lineCount - 1);
-
-  const fullTextRange = new vscode.Range(
-    firstLine.range.start,
-    lastLine.range.end
-  );
-
+): Promise<vscode.TextEdit[]> {
   try {
-    const formatted = prettier.format(source, options);
+    const formatter = vscode.workspace
+      .getConfiguration("hardhat")
+      .get("formatter");
 
-    return [vscode.TextEdit.replace(fullTextRange, formatted)];
+    switch (formatter) {
+      case "prettier":
+        return await prettier.formatDocument(document, context);
+      case "forge":
+        return await forge.formatDocument(document);
+      default:
+        return [];
+    }
   } catch {
-    // ignore formatter errors, we get lots of fails
     return [];
   }
-}
-
-function getCurrentWorkspaceRootFsPath(): string | undefined {
-  const rootFolder = getCurrentWorkspaceRootFolder();
-
-  return rootFolder?.uri.fsPath;
-}
-
-function getCurrentWorkspaceRootFolder(): vscode.WorkspaceFolder | undefined {
-  const editor = vscode.window.activeTextEditor;
-
-  if (!editor) {
-    return undefined;
-  }
-
-  const currentDocument = editor.document.uri;
-
-  return vscode.workspace.getWorkspaceFolder(currentDocument);
-}
-
-function defaultConfig() {
-  return {
-    printWidth: 80,
-    tabWidth: 4,
-    useTabs: false,
-    singleQuote: false,
-    bracketSpacing: false,
-    explicitTypes: "preserve",
-    overrides: [
-      {
-        files: "*.sol",
-        options: {
-          printWidth: 80,
-          tabWidth: 4,
-          useTabs: false,
-          singleQuote: false,
-          bracketSpacing: false,
-          explicitTypes: "preserve",
-        },
-      },
-    ],
-  };
 }
