+++ package.json
@@ -107,6 +107,16 @@
           "type": "boolean",
           "markdownDescription": "Allow **Hardhat for Visual Studio Code** to send extension telemetry. This helps us understand how **Hardhat for Visual Studio Code** is used and how it is performing. Read more in our [privacy policy](https://hardhat.org/privacy-policy.html).\n\n&nbsp;\n\n*__Note:__ **Hardhat for Visual Studio Code** respects the global **Telemetry Level** setting, and will only send telemetry if enabled at both global and extension level.*",
           "default": false
+        },
+        "hardhat.formatter": {
+          "type": "string",
+          "default": "prettier",
+          "enum": [
+            "none",
+            "prettier",
+            "forge"
+          ],
+          "description": "Enables / disables the solidity formatter (prettier solidity default)"
         }
       }
     },
